// frontend/src/lib/api/generated/model/stockItemRead.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Construction ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface StockItemRead {
  /**
     * @minLength 1
     * @maxLength 40
     */
  code: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  category?: string | null;
  /**
     * @minLength 1
     * @maxLength 20
     */
  unit: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  reorder_level?: string;
  notes?: string | null;
  id: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  unit_cost: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  qty_on_hand: string;
  is_active: boolean;
  low_stock?: boolean;
  created_at: string;
}